+++ src/pages/home/components/header.tsx
@@ -1,0 +1,3 @@
+export const Header = (props: { title: string }) => {
+  return <h1 className="header">{props.title}</h1>;
+};