+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 0.3.0 [2021-7-19]
 
-- remove stylus 
+- webpack config使用typescript
+- 删除webpack 5不需要依赖   css-minimizer-webpack-plugin terser-webpack-plugin  clean-webpack-plugin  file-loader url-loader
+- 删除不常用的loader        xml-loader csv-loader  html-loader
+- 删除 stylus 
+- 默认为一个react脚手架，要做js库，需要删除一些依赖和配置
 
 ## 0.2.0 [2020-11-12]